Fixing MAC address
This version of the driver can not get read the mac from
the firmware (uboot). Probably something in the protocol
has changed and the armbiam is not in sync with the
orangpi repository. Until we fix it, lets just create
a MAC address in the driver in order to get the same IP
in the development board every boot for now.